Added old features that I had on my disk.

+++ Source/Foundation/Kernel/Elixir-Application.cpp
@@ -1,5 +1,5 @@
 // -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
-// Copyright (C) 2016-2020 by Agustin Alvarez. All rights reserved.
+// Copyright (C) 2016-2021 by Agustin Alvarez. All rights reserved.
 //
 // This work is licensed under the terms of the Apache License, Version 2.0.
 //
@@ -41,7 +41,7 @@ namespace Elixir::Application
 
     bool Is_Pressed(uint32_t Key)
     {
-        return GetAsyncKeyState(Key) & (1 << 16);
+        return GetAsyncKeyState(static_cast<int>(Key)) & (1 << 16);
     }
 }
